requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
python-dateutil>=2.8.2
openpyxl>=3.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0